+++ task/vsc7448/src/bsp.rs
@@ -6,12 +6,13 @@
 // but you should import Bsp instead, which is autoselected based on board.
 
 cfg_if::cfg_if! {
-    // We use the gemini_bu Bsp for both the Gemini BU and the Gimletlet host.
-    // In both cases, these are driving a VSC7448 dev kit on someone's desk.
+    // We use the vsc7448_dev Bsp for both Gemini-BU and Gimletlet hosts.
+    // In both cases, these are driving a VSC7448 dev kit on someone's desk,
+    // connected over wires to the Hubris board.
     if #[cfg(any(target_board = "gemini-bu-1",
                  target_board = "gimletlet-2"))] {
-        pub mod gemini_bu;
-        pub use gemini_bu::Bsp;
+        pub mod vsc7448_dev;
+        pub use vsc7448_dev::Bsp;
     } else if #[cfg(target_board = "sidecar-1")] {
         pub mod sidecar_1;
         pub use sidecar_1::Bsp;
